training fixed
